refactor(server): retire the per type field routes
