+++ internal/server/server.go
@@ -115,7 +115,6 @@ func NewServer(cfg Config) http.Handler {
 func (s *server) mountOpen(r chi.Router, cfg Config) {
 	if cfg.Types != nil {
 		r.Get("/api/types", s.handleTypeList())
-		r.Get("/api/types/{key}/fields", s.handleFieldList())
 		r.Get("/api/groups", s.handleGroupList())
 		r.Get("/api/groups/params", s.handleGroupParams())
 	}
@@ -157,10 +156,6 @@ func (s *server) mountAdmin(r chi.Router, admin *authkit.AdminHandlers, cfg Conf
 		r.Post("/api/types", s.handleTypeCreate())
 		r.Patch("/api/types/{key}", s.handleTypePatch())
 		r.Delete("/api/types/{key}", s.handleTypeDelete())
-		r.Post("/api/types/{key}/fields", s.handleFieldCreate())
-		r.Put("/api/types/{key}/fields/order", s.handleFieldOrder())
-		r.Patch("/api/types/{key}/fields/{fieldKey}", s.handleFieldPatch())
-		r.Delete("/api/types/{key}/fields/{fieldKey}", s.handleFieldDelete())
 		r.Post("/api/groups", s.handleGroupCreate())
 		r.Put("/api/groups/order", s.handleGroupOrder())
 		r.Patch("/api/groups/{id}", s.handleGroupPatch())
